feat: Upgrade gameBoard factory
This upgrade comes will several features to the game module. Players no
longer need a name and only need a symbol. This is to because symbol
is the only identifier needed.  The debug printing function will output
the winner symbol. We now have a winner symbol that we can keep track
of. getBoard() can return a spot or the whole board. checkWinner() will
only ever declare one winner. resetBoard() has become resetGame() to
reset the entire game instead of just the board.