feat: Add sounds and score
Player objects will now keep track of score. The game will add this
score to each player as needed. The domHandler can take this score and
update it to the DOM.

There is now a sound that plays based on if the click causes a winner
or a regular click if not.