style: Format screen for different sizes
The site will now change based on the size of the screen. This is done
using media queries based on the size, orientation, and changing the
code from using FlexBox randomly everywhere and using CSS Grid because
our layout is pre-defined. The size of the font will now scale magically
using container queries in the CSS.