fix: Various bug fixes
checkWinner will now convert the parameters into numbers to prevent
weird mess ups to the logic happening later. Audio will now play
duplicate sounds because it’s playing a cloned version of the audio.
Game will go back to resetting the screen before starting. It was
removed because the logic was thought to not be needed. It is needed.
The board was set back to being reset to gb.BLANK for consistency even
if using an empty string would be fine.